refactor!: Rename `library-info` to be more intention revealing
Rename `library-info` to `tracking-tag` to be more descriptive and intention revealing.

This change makes the mocking advice added in #448 less confusing as the module name maps to the intent.

BREAKING CHANGE: `"@guardian/cdk/lib/constants/library-info"` renamed to `"@guardian/cdk/lib/constants/tracking-tag"`

Update the mock path in `jest.setup.js` - `"@guardian/cdk/lib/constants/library-info"` becomes `"@guardian/cdk/lib/constants/tracking-tag"`.
