+++ src/constructs/loadbalancing/alb/application-load-balancer.test.ts
@@ -3,7 +3,7 @@ import "../../../utils/test/jest";
 import { SynthUtils } from "@aws-cdk/assert";
 import { Vpc } from "@aws-cdk/aws-ec2";
 import { Stack } from "@aws-cdk/core";
-import { TrackingTag } from "../../../constants/library-info";
+import { TrackingTag } from "../../../constants/tracking-tag";
 import type { SynthedStack } from "../../../utils/test";
 import { alphabeticalTags, simpleGuStackForTesting } from "../../../utils/test";
 import type { AppIdentity } from "../../core/identity";
